+++ muna/muna.py
@@ -42,4 +42,4 @@ class Muna:
         self.users = UserService(self.client)
         self.predictors = PredictorService(self.client)
         self.predictions = PredictionService(self.client)
-        self.beta = BetaClient(self.client, predictions=self.predictions)
+        self.beta = BetaClient(self.client, self.predictors, self.predictions)